feat: Eitherのテストケースを作成
